Convolutional Neural Network Architecture
The PerrosGatosNeural.py script is designed to train a convolutional neural network (CNN) for binary classification of images, specifically to distinguish between cats and dogs. The script uses TensorFlow and Keras libraries to define, train, and evaluate the model. It also includes data augmentation techniques to improve the robustness of the model. (sorry i cant upload all images)